refactor: refactoring to FSD architecture
