+++ shared/components/3dModels/Base.tsx
