+++ shared/components/3dModels/Decoration.tsx
