+++ shared/components/3dModels/Emoji.tsx
