+++ shared/components/3dModels/Ground.tsx
