+++ shared/components/3dModels/MainDecoration.tsx
