+++ shared/components/modals/GuestModal.tsx
@@ -9,8 +9,8 @@ import {
   DialogHeader,
   DialogTitle,
   DialogFooter,
-} from '@/components/ui/dialog';
-import { Button } from '@/components/ui/button';
+} from '@/shared/components/ui/dialog';
+import { Button } from '@/shared/components/ui/button';
 
 const GuestModal = () => {
   const { isOpen, onClose, type } = useModal();
